+++ packages/components/tooltip/src/use-tooltip.ts
@@ -38,10 +38,14 @@ export interface UseTooltipProps
    */
   closeOnClick?: boolean
   /**
-   * If `true`, the tooltip will hide while the mouse
-   * is down
+   * If `true`, the tooltip will hide while the mouse is down
+   * @deprecated - use `closeOnPointerDown` instead
    */
   closeOnMouseDown?: boolean
+  /**
+   * If `true`, the tooltip will hide while the pointer is down
+   */
+  closeOnPointerDown?: boolean
   /**
    * If `true`, the tooltip will hide on pressing Esc key
    * @default true
@@ -87,6 +91,7 @@ export function useTooltip(props: UseTooltipProps = {}) {
     closeDelay = 0,
     closeOnClick = true,
     closeOnMouseDown,
+    closeOnPointerDown = closeOnMouseDown,
     closeOnEsc = true,
     onOpen: onOpenProp,
     onClose: onCloseProp,
@@ -165,11 +170,11 @@ export function useTooltip(props: UseTooltipProps = {}) {
     }
   }, [closeOnClick, closeWithDelay, isOpen])
 
-  const onMouseDown = useCallback(() => {
-    if (isOpen && closeOnMouseDown) {
+  const onPointerDown = useCallback(() => {
+    if (isOpen && closeOnPointerDown) {
       closeWithDelay()
     }
-  }, [closeOnMouseDown, closeWithDelay, isOpen])
+  }, [closeOnPointerDown, closeWithDelay, isOpen])
 
   const onKeyDown = useCallback(
     (event: KeyboardEvent) => {
@@ -195,21 +200,24 @@ export function useTooltip(props: UseTooltipProps = {}) {
   )
 
   /**
-   * This allows for catching mouseleave events when the tooltip
+   * This allows for catching pointerleave events when the tooltip
    * trigger is disabled. There's currently a known issue in
-   * React regarding the onMouseLeave polyfill.
+   * React regarding the onPointerLeave polyfill.
    * @see https://github.com/facebook/react/issues/11972
    */
-  useEventListener(() => ref.current, "mouseleave", closeWithDelay)
+  useEventListener(() => ref.current, "pointerleave", closeWithDelay)
 
   const getTriggerProps: PropGetter = useCallback(
     (props = {}, _ref = null) => {
       const triggerProps = {
         ...props,
         ref: mergeRefs(ref, _ref, referenceRef),
-        onMouseEnter: callAllHandlers(props.onMouseEnter, openWithDelay),
+        onPointerEnter: callAllHandlers(props.onPointerEnter, (e) => {
+          if (e.pointerType === "touch") return
+          openWithDelay()
+        }),
         onClick: callAllHandlers(props.onClick, onClick),
-        onMouseDown: callAllHandlers(props.onMouseDown, onMouseDown),
+        onPointerDown: callAllHandlers(props.onPointerDown, onPointerDown),
         onFocus: callAllHandlers(props.onFocus, openWithDelay),
         onBlur: callAllHandlers(props.onBlur, closeWithDelay),
         "aria-describedby": isOpen ? tooltipId : undefined,
@@ -220,7 +228,7 @@ export function useTooltip(props: UseTooltipProps = {}) {
     [
       openWithDelay,
       closeWithDelay,
-      onMouseDown,
+      onPointerDown,
       isOpen,
       tooltipId,
       onClick,