+++ packages/components/system/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 2.2.9
+
+### Patch Changes
+
+- Updated dependencies
+  [[`ec23eb020`](https://github.com/chakra-ui/chakra-ui/commit/ec23eb02066628d28f85c2c515fba4d0c0120601)]:
+  - @chakra-ui/styled-system@2.3.1
+
 ## 2.2.8
 
 ### Patch Changes